# service/src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  termsOfService: http://www.jelli.com/terms/
  contact:
    email: info@jelli.com
  license:
    name: proprietary
  version: "1.0"
servers:
- url: /api
  variables: {}
...
paths:
...
  /v1/nexus/{repositoryName}/{artifactName}/getReleaseVersions:
    get:
      tags:
      - nexus
      summary: Returns a list of versions for the given artifact name
      operationId: getReleaseVersions
      parameters:
      - name: repositoryName
        in: path
        required: true
        schema:
          type: string
      - name: artifactName
        in: path
        required: true
        schema:
          type: string
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  /v1/nexus/{repositoryName}/{artifactName}/getSnapshotVersions:
    get:
      tags:
      - nexus
      summary: Returns a list of snapshot versions for the given artifact name
      operationId: getSnapshotVersions
      parameters:
      - name: repositoryName
        in: path
        required: true
        schema:
          type: string
      - name: artifactName
        in: path
        required: true
        schema:
          type: string
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
...